fix: green CI after merging main — cover new Inst variants + finish fiber-handle i64
Merging main surfaced two compile breakages in the PR's merge build (the cause
of the red across build/test/windows/macos/loom/svm-llvm):

1. svm-peval: main added Inst::Fma / Inst::VFma / Inst::VDotI8 (scalar/vector
   fused-multiply-add, i8 dot). The exhaustive (by-design, no-wildcard)
   map_operands matched non-exhaustively. Added them: Fma/VFma carry three value
   operands (a,b,c), VDotI8 two (a,b); all are pure single-result, so they also
   join is_removable_if_dead (and thus the specializer's generic residual
   fallback emits them faithfully).

2. svm-jit: PR #68's fiber-handle i32->i64 migration missed the debug backtrace
   path — fiber_handle_generation/resolve take i64 but with_parked_stack and
   fiber_backtrace still took i32, so svm-jit (and everything depending on it)
   failed to compile. Completed the migration: both now take i64 (the only
   callers pass an integer literal, unaffected).

Workspace builds; cargo fmt --all --check and clippy --workspace --all-targets
-D warnings clean; svm-peval (23) and jit_fibers (11) tests pass.

+++ crates/svm-jit/src/lib.rs
@@ -1338,7 +1338,7 @@ impl CompiledModule {
     /// suspend switch. Adjacent duplicate positions are collapsed. Empty if `handle` names no parked
     /// fiber or the module carried no `-g`. Host-side tooling, off the running guest's path (§2a).
     #[cfg(fiber_rt)]
-    pub fn fiber_backtrace(&self, handle: i32) -> Vec<JitFrameLoc> {
+    pub fn fiber_backtrace(&self, handle: i64) -> Vec<JitFrameLoc> {
         let Some(table) = self.fiber_table.as_ref() else {
             return Vec::new();
         };
